more logical location for the predicate
