+++ pkg/vulnerabilityreport/controller/workload.go
@@ -14,8 +14,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/aquasecurity/trivy-operator/pkg/apis/aquasecurity/v1alpha1"
@@ -34,45 +32,6 @@ import (
 )
 
 const trivyServerUp = "trivy_server_up"
-
-// WorkloadPodSpecChangedPredicate creates a predicate that only triggers reconciliation
-// when the workload's podSpec actually changes, not just metadata updates.
-func WorkloadPodSpecChangedPredicate() predicate.Predicate {
-	return predicate.Funcs{
-		UpdateFunc: func(e event.UpdateEvent) bool {
-			if e.ObjectOld == nil || e.ObjectNew == nil {
-				return true // Always process creation/deletion
-			}
-
-			// Extract podSpecs from both old and new objects
-			oldPodSpec, err := kube.GetPodSpec(e.ObjectOld)
-			if err != nil {
-				return true // If we can't get podSpec, process it
-			}
-
-			newPodSpec, err := kube.GetPodSpec(e.ObjectNew)
-			if err != nil {
-				return true // If we can't get podSpec, process it
-			}
-
-			// Compare podSpec hashes
-			oldHash := kube.ComputeHash(oldPodSpec)
-			newHash := kube.ComputeHash(newPodSpec)
-
-			// Only reconcile if podSpec actually changed
-			return oldHash != newHash
-		},
-		CreateFunc: func(e event.CreateEvent) bool {
-			return true // Always process new workloads
-		},
-		DeleteFunc: func(e event.DeleteEvent) bool {
-			return true // Always process deletions
-		},
-		GenericFunc: func(e event.GenericEvent) bool {
-			return true // Process generic events
-		},
-	}
-}
 
 // WorkloadController watches Kubernetes workloads and generates
 // v1alpha1.VulnerabilityReport instances using vulnerability scanner that that
